requirements.txt
click==8.1.3
DateTime==4.7
Flask==2.2.2
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.23.4
pandas==1.5.1
Pillow==9.3.0
python-dateutil==2.8.2
pytz==2022.6
six==1.16.0
Werkzeug==2.2.2
zipp==3.10.0
zope.interface==5.5.1

gunicorn==21.2.0